uri-map and event extension support
This allows the plugin to be used by actual hosts that do not support
atom and urid yet.  I suppose these plugins will probably just
be released as uri-map, but the option to test with urid and atom
have been retained.

NOTE: Compiling with --use-urid or --use-atom results in a plugin
not compatible with the actual specified port types and is therefore
for development purposes only

TODO:
The plugin seems to work well in zynjacku.  I need to finish fixing
the filters, fix dead notes (they should not be dead if no other
note is playing), fix slide toggle (it doesn't toggle off), and
a implement other waveforms.