First sign of life from TripleOscillator.
Basically the VCO works and there is polyphony.  Need to work on
calculating detuning and phase-offset coefficients per-lmms.  Also,
need to replace oscillator 'float' fields to 'float*' or whatever
is appropriate for tracking port changes while a note is sounding.

Then, the VCA, VCF, LFO, and arpeggiator which will be reused for
most other instruments.

Finally, splitting out the voice stealing and event loop code for reuse.
Oh yeah-- need to add a thread to handle reallocating the voices when
polyphony changes.