Almost fully operational LMMS Envelope generator
Completely rewritten.  I really don't see the benefit of precalculating these envelopes.
and if it should become a problem, then yes, we can add a way to do that in a realtime safe way

Room for optimisation, but should behave close to the original.  I preferred sample-accuracy in
some places over 100% emulation.