Properly compile resid, rename sid to sidemu
Here is a wscript to actually compile the resid library.  However, there
was a nasty naming collision due the the "sid.h" file in both resid and
our sid instrument.  Thus, I'm renaming our sid plugin to sidemu (for
SID emulator, as it is seen in the original LMMS artwork.)  Anyone is
welcome to change the name to something cooler.

Also note the sidemu.cc is now able to use ReSID's SID object. Yay!